Port Sudoku Solver to Android
Add UI for Sudoku Solver so that touch screen can be used for input,
numbers only.  Solve will occur when the solve button is pressed.

Make custom Pair object since apparently android does not support the
same pair object that I was using in my plain java code.

Did some color tweaking since I could not seem to figure out how to
display some grid lines without adding a bunch of code, so the colors
are there to help with knowing which square numbers are being entered
in to.

TO DO:
--------
1)Add more graceful response to not finding an answer, right now the
grid (with 0s) is just redisplayed

2)Give the numbers entered (not solved by the program) a highlight so
that they can be distinguished

3)Fix some formatting issues.